+++ tests/forms-schema/forms-schema.test.ts
@@ -4894,26 +4894,32 @@ test('should allow defaultValue for files element with "public" storageType', ()
   expect(value?.defaultValue).toEqual(input.defaultValue)
 })
 
-test('should allow defaultValue for files element with "legacy" storageType', () => {
+test('should not allow invalid values for storageType for files element', () => {
+  const input = {
+    id: 'ff9b04c3-f2ad-4994-a525-e7189eb67a79',
+    name: 'files',
+    label: 'Files',
+    type: 'files',
+    storageType: 'invalid',
+  }
+  const { error } = elementSchema.validate(input)
+  expect(error?.message).toBe('"storageType" must be one of [public, private]')
+})
+
+test('should convert "legacy" to "private" storageType for files element', () => {
   const input = {
     id: 'ff9b04c3-f2ad-4994-a525-e7189eb67a79',
     name: 'files',
     label: 'Files',
     type: 'files',
     storageType: 'legacy',
-    defaultValue: [
-      {
-        fileName: 'dot.png',
-        data: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAMAAAADCAYAAABWKLW/AAAAIElEQVQYV2NkYGBoYGBgqGdgYGhkZGBg+M8ABSAOXAYATFcEA8STCz8AAAAASUVORK5CYII=',
-      },
-    ],
   }
   const { value, error } = elementSchema.validate(input)
   expect(error).toBeUndefined()
-  expect(value?.defaultValue).toEqual(input.defaultValue)
+  expect(value?.storageType).toBe('private')
 })
 
-test('should allow defaultValue in "legacy" format for files element with no storageType', () => {
+test('should allow defaultValue in "private" format for files element with no storageType', () => {
   const input = {
     id: 'ff9b04c3-f2ad-4994-a525-e7189eb67a79',
     name: 'files',
@@ -4921,8 +4927,16 @@ test('should allow defaultValue in "legacy" format for files element with no sto
     type: 'files',
     defaultValue: [
       {
+        s3: {
+          region: 'ap-southeast-2',
+          bucket: 'customer.forms.oneblink.io',
+          key: 'submissions/1/attachments/44cdee6f-edbd-4620-aaf5-df25ce976e43',
+        },
+        url: 'https://auth-api.blinkm.io/submissions/1/attachments/44cdee6f-edbd-4620-aaf5-df25ce976e43',
+        contentType: 'image/png',
         fileName: 'dot.png',
-        data: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAMAAAADCAYAAABWKLW/AAAAIElEQVQYV2NkYGBoYGBgqGdgYGhkZGBg+M8ABSAOXAYATFcEA8STCz8AAAAASUVORK5CYII=',
+        id: '44cdee6f-edbd-4620-aaf5-df25ce976e43',
+        isPrivate: true,
       },
     ],
   }
@@ -4981,29 +4995,37 @@ test('should allow defaultValue for camera element with "public" storageType', (
   expect(value?.defaultValue).toEqual(input.defaultValue)
 })
 
-test('should allow defaultValue for camera element with "legacy" storageType', () => {
+test('should convert "legacy" to "private" storageType for camera element', () => {
   const input = {
     id: 'ff9b04c3-f2ad-4994-a525-e7189eb67a79',
     name: 'camera',
     label: 'Camera',
     type: 'camera',
     storageType: 'legacy',
-    defaultValue:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAMAAAADCAYAAABWKLW/AAAAIElEQVQYV2NkYGBoYGBgqGdgYGhkZGBg+M8ABSAOXAYATFcEA8STCz8AAAAASUVORK5CYII=',
   }
   const { value, error } = elementSchema.validate(input)
   expect(error).toBeUndefined()
-  expect(value?.defaultValue).toEqual(input.defaultValue)
+  expect(value?.storageType).toBe('private')
 })
 
-test('should allow defaultValue in "legacy" format for camera element with no storageType', () => {
+test('should allow defaultValue in "private" format for files element with no storageType', () => {
   const input = {
     id: 'ff9b04c3-f2ad-4994-a525-e7189eb67a79',
     name: 'camera',
     label: 'Camera',
     type: 'camera',
-    defaultValue:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAMAAAADCAYAAABWKLW/AAAAIElEQVQYV2NkYGBoYGBgqGdgYGhkZGBg+M8ABSAOXAYATFcEA8STCz8AAAAASUVORK5CYII=',
+    defaultValue: {
+      s3: {
+        region: 'ap-southeast-2',
+        bucket: 'customer.forms.oneblink.io',
+        key: 'submissions/1/attachments/44cdee6f-edbd-4620-aaf5-df25ce976e43',
+      },
+      url: 'https://auth-api.blinkm.io/submissions/1/attachments/44cdee6f-edbd-4620-aaf5-df25ce976e43',
+      contentType: 'image/png',
+      fileName: 'dot.png',
+      id: '44cdee6f-edbd-4620-aaf5-df25ce976e43',
+      isPrivate: true,
+    },
   }
   const { value, error } = elementSchema.validate(input)
   expect(error).toBeUndefined()
@@ -5060,29 +5082,37 @@ test('should allow defaultValue for draw element with "public" storageType', () 
   expect(value?.defaultValue).toEqual(input.defaultValue)
 })
 
-test('should allow defaultValue for draw element with "legacy" storageType', () => {
+test('should convert "legacy" to "private" storageType for draw element', () => {
   const input = {
     id: 'ff9b04c3-f2ad-4994-a525-e7189eb67a79',
     name: 'draw',
     label: 'Draw',
     type: 'draw',
     storageType: 'legacy',
-    defaultValue:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAMAAAADCAYAAABWKLW/AAAAIElEQVQYV2NkYGBoYGBgqGdgYGhkZGBg+M8ABSAOXAYATFcEA8STCz8AAAAASUVORK5CYII=',
   }
   const { value, error } = elementSchema.validate(input)
   expect(error).toBeUndefined()
-  expect(value?.defaultValue).toEqual(input.defaultValue)
+  expect(value?.storageType).toBe('private')
 })
 
-test('should allow defaultValue in "legacy" format for draw element with no storageType', () => {
+test('should allow defaultValue in "private" format for draw element with no storageType', () => {
   const input = {
     id: 'ff9b04c3-f2ad-4994-a525-e7189eb67a79',
     name: 'draw',
     label: 'Draw',
     type: 'draw',
-    defaultValue:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAMAAAADCAYAAABWKLW/AAAAIElEQVQYV2NkYGBoYGBgqGdgYGhkZGBg+M8ABSAOXAYATFcEA8STCz8AAAAASUVORK5CYII=',
+    defaultValue: {
+      s3: {
+        region: 'ap-southeast-2',
+        bucket: 'customer.forms.oneblink.io',
+        key: 'submissions/1/attachments/44cdee6f-edbd-4620-aaf5-df25ce976e43',
+      },
+      url: 'https://auth-api.blinkm.io/submissions/1/attachments/44cdee6f-edbd-4620-aaf5-df25ce976e43',
+      contentType: 'image/png',
+      fileName: 'dot.png',
+      id: '44cdee6f-edbd-4620-aaf5-df25ce976e43',
+      isPrivate: true,
+    },
   }
   const { value, error } = elementSchema.validate(input)
   expect(error).toBeUndefined()